fix: size of the Feedback object
It seems that the size changed along firmware version (problem we had with other
objects)

+++ src/Objects.hpp
@@ -22,7 +22,7 @@ namespace motors_roboteq_canopen {
     CANOPEN_DEFINE_OBJECT(0x210F, 1, TemperatureMCU,                std::int16_t);
     CANOPEN_DEFINE_OBJECT(0x210F, 2, TemperatureSensor0,            std::int16_t);
 
-    CANOPEN_DEFINE_OBJECT(0x2110, 1, Feedback,                      std::int16_t);
+    CANOPEN_DEFINE_OBJECT(0x2110, 1, Feedback,                      std::int32_t);
     CANOPEN_DEFINE_OBJECT(0x2111, 0, StatusFlagsRaw,                std::uint16_t);
     CANOPEN_DEFINE_OBJECT(0x2112, 0, FaultFlagsRaw,                 std::uint16_t);
     CANOPEN_DEFINE_OBJECT(0x2113, 0, ReadAllDigitalOutput,          std::uint16_t);
